Add files via upload
This is IOT Project of my Master of Computer Application Semester-2 Assignment, where I build with my team members a RFID Student Attendance System. In which student which time will come in the classroom and go that's this project through we will check and attendance that the student in the lecture in the classroom or not.